chore(FDS-349): ViewConfig default icon to eye
