+++ .changeset/cuddly-mugs-exist.md
@@ -1,0 +1,5 @@
+---
+'@espressive/cascara': patch
+---
+
+**ViewConfig:** Change default trigger icon, memoized options list, loading state
