+++ .changeset/fuzzy-pets-destroy.md
@@ -1,0 +1,5 @@
+---
+'@espressive/icons': minor
+---
+
+**NEW:** eye icon
