+++ packages/cascara/src/private/ViewConfig/__globals.js
@@ -2,11 +2,17 @@ import React from 'react';
 import pt from 'prop-types';
 import { Button } from 'reakit/Button';
 import { InlineIcon } from '@iconify/react';
-import { verticalmenuIcon } from '@espressive/icons';
+import { eyeIcon } from '@espressive/icons';
 
 export const DEFAULT_TRIGGER = (
   <Button className='ui basic icon button'>
-    <InlineIcon icon={verticalmenuIcon} />
+    <InlineIcon icon={eyeIcon} />
+  </Button>
+);
+
+export const LOADING_TRIGGER = (
+  <Button className='ui basic loading icon button' disabled>
+    <InlineIcon icon={eyeIcon} />
   </Button>
 );
 
@@ -17,3 +23,12 @@ export const STATE_SHAPE = pt.exact({
   moveItemDown: pt.func.isRequired,
   moveItemUp: pt.func.isRequired,
 });
+
+export const VIEW_CONFIG_PROP_TYPES = {
+  isInitialOpen: pt.bool,
+  options: pt.arrayOf(pt.object).isRequired,
+  placement: pt.oneOf(['bottom-end', 'bottom-start']),
+  state: STATE_SHAPE,
+  title: pt.string,
+  trigger: pt.node,
+};
