+++ packages/cascara/src/private/ViewConfig/index.js
@@ -1,2 +1,3 @@
 export { default as useViewConfigState } from './useViewConfigState';
+export { VIEW_CONFIG_PROP_TYPES as viewConfigPropTypes } from './__globals';
 export { default } from './ViewConfig';
